Fixed 🐞 in Delete Post And CommentSection

+++ app/backend/src/controllers/PostController.js
@@ -4,7 +4,7 @@ const asyncHandler = require("express-async-handler");
 const path = require('path');
 
 
-const PostUpload = async (req, res) => {
+const CreatePost = async (req, res) => {
     const { title, content } = req.body;
     const user_id = req.user.id;
     console.log(user_id);
@@ -37,7 +37,7 @@ const PostUpload = async (req, res) => {
                     });
             });
 
-            res.status(201);
+            res.status(201).json({ message: "Post Uploaded Succesfully" });
         }
         else {
             res.status(400);
@@ -53,10 +53,11 @@ const PostUpload = async (req, res) => {
 }
 const GetPost = async (req, res) => {
     try {
-        const userId = req.params.userId;
+        const userId = req.user.id;
         const userPosts = await Post.find({ userid: userId });
 
-        res.render('userPosts', { posts: userPosts });
+
+        res.json({ posts: userPosts });
     } catch (error) {
         console.error(error);
         res.status(500).json({ error: 'Internal Server Error' });
@@ -67,18 +68,25 @@ const GetPost = async (req, res) => {
 const DeletePost = async (req, res) => {
     try {
         const userId = req.user.id;
-        const postId = req.params.postid;
-        const user = await User.findById(userId);
+        const postId = req.params._id;
+        if (!postId) {
 
+            return res.status(404).json({ error: 'Post not found' });
+        }
+
+
+        // Check if the user exists
+        const user = await User.findById(userId);
         if (!user) {
             return res.status(404).json({ error: 'User not found' });
         }
 
+        // Remove the postId from the user's posts array
         user.posts.pull(postId);
         await user.save();
 
+        // Delete the post by its postId
         await Post.deleteOne({ _id: postId });
-        await Post.save();
 
         res.status(200).json({ message: 'Post deleted successfully' });
     } catch (error) {
@@ -87,9 +95,64 @@ const DeletePost = async (req, res) => {
     }
 };
 
+const PostComment = async (req, res) => {
+    const { comment } = req.body;
 
+    if (!comment) {
+        res.status(400).json({ error: "All fields are mandatory" });
+        return;
+    }
+
+    try {
+        const user = await User.findById(req.user.id);
+        const PostId = req.params._id;
+        if (!PostId) {
+
+            return res.status(404).json({ error: 'Post not found' });
+        }
+        const post = await Post.findById(PostId);
+        console.log("Post:", post);
+
+
+        const newComment = {
+            comment: req.body.comment,
+            userid: req.user.id,
+            name: user.name
+        };
+
+        // Use $push to add the new comment to the end of the 'comments' array
+        await Post.updateOne({ _id: post }, { $push: { comments: newComment } });
+        // await Post.save();
+
+        res.status(200).json({ message: "Comment Added Succesfully" });
+    } catch (error) {
+        console.error(error);
+        res.status(500).json({ error: "Internal Server Error" });
+    }
+};
+
+const GetComment = async (req, res) => {
+    try {
+
+        const postId = req.params._id;
+        if (!postId) {
+
+            return res.status(404).json({ error: 'Post not found' });
+        }
+        // Assuming you get the postId from the request parameters
+        const post = await Post.findById(postId).select('comments');
+
+        res.json({ comments: post.comments });
+
+    } catch (error) {
+        console.error(error);
+        res.status(500).json({ error: 'Internal Server Error' });
+    }
+}
 module.exports = {
-    PostUpload,
+    CreatePost,
     GetPost,
-    DeletePost
+    DeletePost,
+    PostComment,
+    GetComment
 }
